… crates/ide-diagnostics/src/lib.rs
… crates/rust-analyzer/src/cli/analysis_stats.rs
… crates/rust-analyzer/src/config.rs
… crates/rust-analyzer/src/integrated_benchmarks.rs
… docs/user/generated_config.adoc
… editors/code/package.json